+++ packages/bluesky/lib/src/services/entities/converter/embed_view_record_view_converter.dart
@@ -8,11 +8,13 @@ import 'package:freezed_annotation/freezed_annotation.dart';
 
 // 🌎 Project imports:
 import '../../../ids.g.dart' as ids;
+import '../../types/app/bsky/embed/record/embed_record_view_detached.dart';
 import '../embed_view_record_view.dart';
 import '../embed_view_record_view_blocked.dart';
 import '../embed_view_record_view_not_found.dart';
 import '../embed_view_record_view_record.dart';
 import '../feed_generator_view.dart';
+import '../labeler_view.dart';
 import '../list_view.dart';
 
 const embedViewRecordViewConverter = _EmbedViewRecordViewConverter();
@@ -39,6 +41,10 @@ final class _EmbedViewRecordViewConverter
         return EmbedViewRecordView.blocked(
           data: EmbedViewRecordViewBlocked.fromJson(json),
         );
+      } else if (type == ids.appBskyEmbedRecordViewDetached) {
+        return EmbedViewRecordView.viewDetached(
+          data: EmbedRecordViewDetached.fromJson(json),
+        );
       } else if (type == ids.appBskyFeedDefsGeneratorView) {
         return EmbedViewRecordView.generatorView(
           data: FeedGeneratorView.fromJson(json),
@@ -46,6 +52,10 @@ final class _EmbedViewRecordViewConverter
       } else if (type == ids.appBskyGraphDefsListView) {
         return EmbedViewRecordView.listView(
           data: ListView.fromJson(json),
+        );
+      } else if (type == ids.appBskyLabelerDefsLabelerView) {
+        return EmbedViewRecordView.labelerView(
+          data: LabelerView.fromJson(json),
         );
       }
 
@@ -60,8 +70,10 @@ final class _EmbedViewRecordViewConverter
         record: (data) => data.toJson(),
         notFound: (data) => data.toJson(),
         blocked: (data) => data.toJson(),
+        viewDetached: (data) => data.toJson(),
         generatorView: (data) => data.toJson(),
         listView: (data) => data.toJson(),
+        labelerView: (data) => data.toJson(),
         unknown: (data) => data,
       );
 }