+++ packages/bluesky/lib/src/services/types/app/bsky/feed/postgate/embedding_rules.freezed.dart
@@ -1,0 +1,408 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'embedding_rules.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#adding-getters-and-methods-to-our-models');
+
+/// @nodoc
+mixin _$UPostgateEmbeddingRule {
+  Object get data => throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function(PostgateDisableRule data) disableRule,
+    required TResult Function(Map<String, dynamic> data) unknown,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function(PostgateDisableRule data)? disableRule,
+    TResult? Function(Map<String, dynamic> data)? unknown,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(PostgateDisableRule data)? disableRule,
+    TResult Function(Map<String, dynamic> data)? unknown,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(UPostgateEmbeddingRuleDisableRule value)
+        disableRule,
+    required TResult Function(UPostgateEmbeddingRuleUnknown value) unknown,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(UPostgateEmbeddingRuleDisableRule value)? disableRule,
+    TResult? Function(UPostgateEmbeddingRuleUnknown value)? unknown,
+  }) =>
+      throw _privateConstructorUsedError;
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(UPostgateEmbeddingRuleDisableRule value)? disableRule,
+    TResult Function(UPostgateEmbeddingRuleUnknown value)? unknown,
+    required TResult orElse(),
+  }) =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $UPostgateEmbeddingRuleCopyWith<$Res> {
+  factory $UPostgateEmbeddingRuleCopyWith(UPostgateEmbeddingRule value,
+          $Res Function(UPostgateEmbeddingRule) then) =
+      _$UPostgateEmbeddingRuleCopyWithImpl<$Res, UPostgateEmbeddingRule>;
+}
+
+/// @nodoc
+class _$UPostgateEmbeddingRuleCopyWithImpl<$Res,
+        $Val extends UPostgateEmbeddingRule>
+    implements $UPostgateEmbeddingRuleCopyWith<$Res> {
+  _$UPostgateEmbeddingRuleCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  /// Create a copy of UPostgateEmbeddingRule
+  /// with the given fields replaced by the non-null parameter values.
+}
+
+/// @nodoc
+abstract class _$$UPostgateEmbeddingRuleDisableRuleImplCopyWith<$Res> {
+  factory _$$UPostgateEmbeddingRuleDisableRuleImplCopyWith(
+          _$UPostgateEmbeddingRuleDisableRuleImpl value,
+          $Res Function(_$UPostgateEmbeddingRuleDisableRuleImpl) then) =
+      __$$UPostgateEmbeddingRuleDisableRuleImplCopyWithImpl<$Res>;
+  @useResult
+  $Res call({PostgateDisableRule data});
+
+  $PostgateDisableRuleCopyWith<$Res> get data;
+}
+
+/// @nodoc
+class __$$UPostgateEmbeddingRuleDisableRuleImplCopyWithImpl<$Res>
+    extends _$UPostgateEmbeddingRuleCopyWithImpl<$Res,
+        _$UPostgateEmbeddingRuleDisableRuleImpl>
+    implements _$$UPostgateEmbeddingRuleDisableRuleImplCopyWith<$Res> {
+  __$$UPostgateEmbeddingRuleDisableRuleImplCopyWithImpl(
+      _$UPostgateEmbeddingRuleDisableRuleImpl _value,
+      $Res Function(_$UPostgateEmbeddingRuleDisableRuleImpl) _then)
+      : super(_value, _then);
+
+  /// Create a copy of UPostgateEmbeddingRule
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? data = null,
+  }) {
+    return _then(_$UPostgateEmbeddingRuleDisableRuleImpl(
+      data: null == data
+          ? _value.data
+          : data // ignore: cast_nullable_to_non_nullable
+              as PostgateDisableRule,
+    ));
+  }
+
+  /// Create a copy of UPostgateEmbeddingRule
+  /// with the given fields replaced by the non-null parameter values.
+  @override
+  @pragma('vm:prefer-inline')
+  $PostgateDisableRuleCopyWith<$Res> get data {
+    return $PostgateDisableRuleCopyWith<$Res>(_value.data, (value) {
+      return _then(_value.copyWith(data: value));
+    });
+  }
+}
+
+/// @nodoc
+
+class _$UPostgateEmbeddingRuleDisableRuleImpl
+    extends UPostgateEmbeddingRuleDisableRule {
+  const _$UPostgateEmbeddingRuleDisableRuleImpl({required this.data})
+      : super._();
+
+  @override
+  final PostgateDisableRule data;
+
+  @override
+  String toString() {
+    return 'UPostgateEmbeddingRule.disableRule(data: $data)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$UPostgateEmbeddingRuleDisableRuleImpl &&
+            (identical(other.data, data) || other.data == data));
+  }
+
+  @override
+  int get hashCode => Object.hash(runtimeType, data);
+
+  /// Create a copy of UPostgateEmbeddingRule
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$UPostgateEmbeddingRuleDisableRuleImplCopyWith<
+          _$UPostgateEmbeddingRuleDisableRuleImpl>
+      get copyWith => __$$UPostgateEmbeddingRuleDisableRuleImplCopyWithImpl<
+          _$UPostgateEmbeddingRuleDisableRuleImpl>(this, _$identity);
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function(PostgateDisableRule data) disableRule,
+    required TResult Function(Map<String, dynamic> data) unknown,
+  }) {
+    return disableRule(data);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function(PostgateDisableRule data)? disableRule,
+    TResult? Function(Map<String, dynamic> data)? unknown,
+  }) {
+    return disableRule?.call(data);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(PostgateDisableRule data)? disableRule,
+    TResult Function(Map<String, dynamic> data)? unknown,
+    required TResult orElse(),
+  }) {
+    if (disableRule != null) {
+      return disableRule(data);
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(UPostgateEmbeddingRuleDisableRule value)
+        disableRule,
+    required TResult Function(UPostgateEmbeddingRuleUnknown value) unknown,
+  }) {
+    return disableRule(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(UPostgateEmbeddingRuleDisableRule value)? disableRule,
+    TResult? Function(UPostgateEmbeddingRuleUnknown value)? unknown,
+  }) {
+    return disableRule?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(UPostgateEmbeddingRuleDisableRule value)? disableRule,
+    TResult Function(UPostgateEmbeddingRuleUnknown value)? unknown,
+    required TResult orElse(),
+  }) {
+    if (disableRule != null) {
+      return disableRule(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class UPostgateEmbeddingRuleDisableRule
+    extends UPostgateEmbeddingRule {
+  const factory UPostgateEmbeddingRuleDisableRule(
+          {required final PostgateDisableRule data}) =
+      _$UPostgateEmbeddingRuleDisableRuleImpl;
+  const UPostgateEmbeddingRuleDisableRule._() : super._();
+
+  @override
+  PostgateDisableRule get data;
+
+  /// Create a copy of UPostgateEmbeddingRule
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  _$$UPostgateEmbeddingRuleDisableRuleImplCopyWith<
+          _$UPostgateEmbeddingRuleDisableRuleImpl>
+      get copyWith => throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class _$$UPostgateEmbeddingRuleUnknownImplCopyWith<$Res> {
+  factory _$$UPostgateEmbeddingRuleUnknownImplCopyWith(
+          _$UPostgateEmbeddingRuleUnknownImpl value,
+          $Res Function(_$UPostgateEmbeddingRuleUnknownImpl) then) =
+      __$$UPostgateEmbeddingRuleUnknownImplCopyWithImpl<$Res>;
+  @useResult
+  $Res call({Map<String, dynamic> data});
+}
+
+/// @nodoc
+class __$$UPostgateEmbeddingRuleUnknownImplCopyWithImpl<$Res>
+    extends _$UPostgateEmbeddingRuleCopyWithImpl<$Res,
+        _$UPostgateEmbeddingRuleUnknownImpl>
+    implements _$$UPostgateEmbeddingRuleUnknownImplCopyWith<$Res> {
+  __$$UPostgateEmbeddingRuleUnknownImplCopyWithImpl(
+      _$UPostgateEmbeddingRuleUnknownImpl _value,
+      $Res Function(_$UPostgateEmbeddingRuleUnknownImpl) _then)
+      : super(_value, _then);
+
+  /// Create a copy of UPostgateEmbeddingRule
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? data = null,
+  }) {
+    return _then(_$UPostgateEmbeddingRuleUnknownImpl(
+      data: null == data
+          ? _value._data
+          : data // ignore: cast_nullable_to_non_nullable
+              as Map<String, dynamic>,
+    ));
+  }
+}
+
+/// @nodoc
+
+class _$UPostgateEmbeddingRuleUnknownImpl
+    extends UPostgateEmbeddingRuleUnknown {
+  const _$UPostgateEmbeddingRuleUnknownImpl(
+      {required final Map<String, dynamic> data})
+      : _data = data,
+        super._();
+
+  final Map<String, dynamic> _data;
+  @override
+  Map<String, dynamic> get data {
+    if (_data is EqualUnmodifiableMapView) return _data;
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableMapView(_data);
+  }
+
+  @override
+  String toString() {
+    return 'UPostgateEmbeddingRule.unknown(data: $data)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$UPostgateEmbeddingRuleUnknownImpl &&
+            const DeepCollectionEquality().equals(other._data, _data));
+  }
+
+  @override
+  int get hashCode =>
+      Object.hash(runtimeType, const DeepCollectionEquality().hash(_data));
+
+  /// Create a copy of UPostgateEmbeddingRule
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$UPostgateEmbeddingRuleUnknownImplCopyWith<
+          _$UPostgateEmbeddingRuleUnknownImpl>
+      get copyWith => __$$UPostgateEmbeddingRuleUnknownImplCopyWithImpl<
+          _$UPostgateEmbeddingRuleUnknownImpl>(this, _$identity);
+
+  @override
+  @optionalTypeArgs
+  TResult when<TResult extends Object?>({
+    required TResult Function(PostgateDisableRule data) disableRule,
+    required TResult Function(Map<String, dynamic> data) unknown,
+  }) {
+    return unknown(data);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? whenOrNull<TResult extends Object?>({
+    TResult? Function(PostgateDisableRule data)? disableRule,
+    TResult? Function(Map<String, dynamic> data)? unknown,
+  }) {
+    return unknown?.call(data);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeWhen<TResult extends Object?>({
+    TResult Function(PostgateDisableRule data)? disableRule,
+    TResult Function(Map<String, dynamic> data)? unknown,
+    required TResult orElse(),
+  }) {
+    if (unknown != null) {
+      return unknown(data);
+    }
+    return orElse();
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult map<TResult extends Object?>({
+    required TResult Function(UPostgateEmbeddingRuleDisableRule value)
+        disableRule,
+    required TResult Function(UPostgateEmbeddingRuleUnknown value) unknown,
+  }) {
+    return unknown(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult? mapOrNull<TResult extends Object?>({
+    TResult? Function(UPostgateEmbeddingRuleDisableRule value)? disableRule,
+    TResult? Function(UPostgateEmbeddingRuleUnknown value)? unknown,
+  }) {
+    return unknown?.call(this);
+  }
+
+  @override
+  @optionalTypeArgs
+  TResult maybeMap<TResult extends Object?>({
+    TResult Function(UPostgateEmbeddingRuleDisableRule value)? disableRule,
+    TResult Function(UPostgateEmbeddingRuleUnknown value)? unknown,
+    required TResult orElse(),
+  }) {
+    if (unknown != null) {
+      return unknown(this);
+    }
+    return orElse();
+  }
+}
+
+abstract class UPostgateEmbeddingRuleUnknown extends UPostgateEmbeddingRule {
+  const factory UPostgateEmbeddingRuleUnknown(
+          {required final Map<String, dynamic> data}) =
+      _$UPostgateEmbeddingRuleUnknownImpl;
+  const UPostgateEmbeddingRuleUnknown._() : super._();
+
+  @override
+  Map<String, dynamic> get data;
+
+  /// Create a copy of UPostgateEmbeddingRule
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  _$$UPostgateEmbeddingRuleUnknownImplCopyWith<
+          _$UPostgateEmbeddingRuleUnknownImpl>
+      get copyWith => throw _privateConstructorUsedError;
+}